+++ src/tasks/mocha/__tests__/source-map-support/gulpfile.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const gulp = require('gulp');
+const task = require('../../index');
+
+task.dependencies = {};
+task.register(gulp, 'mocha', {
+  files: 'index-test.ts'
+});