+++ src/tasks/mocha/__tests__/source-map-support/index-test.ts
@@ -1,0 +1,5 @@
+describe('suite', () => {
+  it('case', async() => {
+    throw new Error('error');
+  });
+});