+++ src/tasks/mocha/defaults.js
@@ -5,5 +5,5 @@ const processCwd = process.cwd();
 
 module.exports = {
   files: path.join(processCwd, 'src/test/**/*.+(js|ts|tsx)'),
-  require: `${path.join(__dirname, 'preprocessor.js')}`
+  require: `${path.join(__dirname, 'setup-file.js')}`
 };